Revert "fix ttwars send farmlist"
This reverts commit 0a1580973d2d94c97fa11f940f58cddc51434473.
